feat: Moved tracing to tuillez
