+++ alistral_core/src/cli/mod.rs
@@ -1,4 +1,3 @@
-pub mod pg_styles;
 pub mod colors;
 pub mod constants;
 pub mod logger;
