+++ alistral_core/src/database/fetching/releases.rs
@@ -1,10 +1,13 @@
 use itertools::Itertools as _;
 use musicbrainz_db_lite::models::musicbrainz::release::Release;
-
-use crate::cli::logger::println_cli;
-use crate::cli::progress_bar::global_progress_bar::PG_FETCHING;
+use tracing::info;
+use tracing::instrument;
+use tracing::Span;
+use tracing_indicatif::span_ext::IndicatifSpanExt as _;
+use tuillez::pg_counted;
 
 /// Prefetch all the release of a list of recordings
+#[instrument(skip(client), fields(indicatif.pb_show = tracing::field::Empty))]
 pub async fn prefetch_releases(
     conn: &mut sqlx::SqliteConnection,
     client: &crate::AlistralClient,
@@ -16,14 +19,14 @@ pub async fn prefetch_releases(
         .filter(|r| !r.is_fully_fetched())
         .collect_vec();
 
-    let progress_bar = PG_FETCHING.get_submitter(uncompletes.len() as u64);
+    pg_counted!(uncompletes.len(), "Fetching releases");
+    info!("Fetching full release data");
 
-    println_cli("Fetching missing release data");
     for release in uncompletes {
         release
             .fetch_if_incomplete(conn, &client.musicbrainz_db)
             .await?;
-        progress_bar.inc(1);
+        Span::current().pb_inc(1);
     }
 
     Ok(())
