+++ alistral_core/src/datastructures/entity_with_listens/artist/collection/artist_with_recordings.rs
@@ -1,6 +1,8 @@
 use itertools::Itertools as _;
 use musicbrainz_db_lite::models::musicbrainz::artist::Artist;
 use musicbrainz_db_lite::models::musicbrainz::recording::Recording;
+use tracing::instrument;
+use tuillez::pg_spinner;
 
 use crate::database::fetching::recordings::fetch_recordings_as_complete;
 use crate::datastructures::entity_with_listens::artist::artist_with_recordings::ArtistWithRecordings;
@@ -12,12 +14,14 @@ pub type ArtistWithRecordingsCollection =
     EntityWithListensCollection<Artist, RecordingWithListensCollection>;
 
 impl ArtistWithRecordingsCollection {
+    #[instrument(skip_all, fields(indicatif.pb_show = tracing::field::Empty))]
     pub async fn from_listencollection(
         conn: &mut sqlx::SqliteConnection,
         client: &crate::AlistralClient,
         listens: ListenCollection,
     ) -> Result<Self, crate::Error> {
         // Convert Recordings
+        pg_spinner!("Compiling artist data");
         let recordings =
             RecordingWithListensCollection::from_listencollection(conn, client, listens).await?;
