+++ alistral_core/src/datastructures/entity_with_listens/release_group/collection/mod.rs
@@ -1,6 +1,8 @@
 use itertools::Itertools as _;
 use musicbrainz_db_lite::models::musicbrainz::release::Release;
 use musicbrainz_db_lite::models::musicbrainz::release_group::ReleaseGroup;
+use tracing::instrument;
+use tuillez::pg_spinner;
 
 use crate::database::fetching::releases::prefetch_releases;
 use crate::datastructures::entity_with_listens::collection::EntityWithListensCollection;
@@ -12,11 +14,13 @@ pub type ReleaseGroupWithListensCollection =
     EntityWithListensCollection<ReleaseGroup, ListenCollection>;
 
 impl ReleaseGroupWithListensCollection {
+    #[instrument(skip_all, fields(indicatif.pb_show = tracing::field::Empty))]
     pub async fn from_listencollection(
         conn: &mut sqlx::SqliteConnection,
         client: &crate::AlistralClient,
         listens: ListenCollection,
     ) -> Result<Self, crate::Error> {
+        pg_spinner!("Compiling release group listens data");
         // Convert Releases
         let releases =
             ReleaseWithListensCollection::from_listencollection(conn, client, listens).await?;
