+++ tuillez/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "tuillez"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+indicatif = "0.17.11"
+owo-colors = "4.1.0"
+tracing-indicatif = "0.3.9"
+
+[lints]
+workspace = true
