+++ tuillez/src/lib.rs
@@ -1,0 +1,4 @@
+pub mod macros;
+pub mod styles;
+
+pub use crate::styles::SPINNER_STYLE;
